Update iframeheight.js
There is surely something wrong with this line. Also the fact that you are forced to have a file name on the end of the URL unnecessarily restrictive. If no name can be resolved for the window is not resize there for its imperative that the expression is able to resolve some name even if its not perfectly formed. I hope you can see what I mean.